pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-agent-os"
version = "0.1.0"
description = "MVP for assisted YouTube video creation and publishing."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "alembic>=1.14.0,<2.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "redis>=5.2.0,<6.0.0",
  "celery>=5.4.0,<6.0.0",
  "httpx>=0.28.0,<1.0.0",
  "python-multipart>=0.0.17,<1.0.0",
  "google-api-python-client>=2.154.0,<3.0.0",
  "google-auth>=2.36.0,<3.0.0",
  "google-auth-oauthlib>=1.2.1,<2.0.0",
  "cryptography>=44.0.0,<47.0.0",
  "agno>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "mypy>=1.13.0,<2.0.0",
  "ruff>=0.8.0,<1.0.0",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
  "N",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q -p no:cacheprovider"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
check_untyped_defs = true
plugins = []
exclude = ["alembic"]

[[tool.mypy.overrides]]
module = ["celery"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["google_auth_oauthlib.*", "googleapiclient.*"]
ignore_missing_imports = true
